Split components into mulitple of single file
Under component folder, each file stands for one purposes
