+++ src/components/Button.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function Button({ children }) {
+  return <button type="button">{children}</button>;
+}
+
+export default Button;
