+++ src/components/Buttons.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import Button from './Button';
+
+function Buttons() {
+  return (
+    <p>
+      {[1, 2, 3].map((i) => <Button key={i}>{i}</Button>)}
+    </p>
+  );
+}
+
+export default Buttons;
